requirements.txt
streamlit
requests
wikipedia
python-dotenv
scikit-learn
pandas
google-api-python-client
matplotlib
pillow